requirements.txt
numpy
pandas
sklearn
flask
flask-cors
gunicorn
PyYAML
dill
pymongo[srv]
dnspython
xgboost
optuna
evidently
atomicwrites
attrs
certifi
click
colorama
dnspython
Flask
Flask-Cors
gunicorn
importlib-metadata
iniconfig
itsdangerous
Jinja2
joblib
MarkupSafe
packaging
pluggy
py
pymongo
pyparsing
pytest
python-dateutil
pytz
scipy
six
sklearn
threadpoolctl
tomli
typing_extensions
Werkzeug
wincertstore
zipp
gunicorn
statsmodels
prophet
pmdarima
lightgbm
streamlit
google-api-python-client
-e .